+++ uefi/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `proto::media::block::BlockIO2`.
 - Added `proto::device_path::DevicePath::to_pool()`.
 - Added `proto::device_path::DevicePathUtilities::duplicate_path()`.
+- Added `proto::pci::enumeration::PciTree::device_path()`.
 
 ## Changed
 - Changed ordering of `proto::pci::PciIoAddress` to (bus -> dev -> fun -> reg -> ext_reg).